add useSnowbridgeContext hook
